+++ packages/codistica-react-icons/src/index.js
@@ -1,0 +1,14 @@
+/** @todo FOLLOW https://github.com/facebook/flow/issues/8354 and https://youtrack.jetbrains.com/issue/WEB-45239 issues. */
+
+// COMPONENTS
+export {Facebook} from './components/facebook';
+export {Instagram} from './components/instagram';
+export {Linkedin} from './components/linkedin';
+export {Twitter} from './components/twitter';
+export {Youtube} from './components/youtube';
+
+// UTILS
+export {Icon} from './utils/icon';
+
+// TODO: ADD GitHub, NPM, BitBucket, etc... TO icons
+// TODO: FIX INSTAGRAM ICON IN ILLUSTRATOR! (CHECK ALL)