+++ packages/codistica-react-icons/stories/components/instagram.stories.js
@@ -2,21 +2,21 @@
 
 import {default as centered} from '@storybook/addon-centered/react';
 import React from 'react';
-import {BGS_DARK} from '../../../.storybook/custom-backgrounds.js';
-import {icons} from '../../../src/index.js';
+import {BGS_DARK} from '../../.storybook/custom-backgrounds.js';
+import {Instagram as InstagramIcon} from '../../src/index.js';
 
 /**
  * @description A Instagram icon demo.
  * @returns {Object<string,*>} React component.
  */
 function Instagram() {
-    return <icons.Instagram href={'https://www.instagram.com'} />;
+    return <InstagramIcon href={'https://www.instagram.com'} />;
 }
 
 export {Instagram};
 
 export default {
-    title: 'Icons',
+    title: 'Socials',
     parameters: {
         backgrounds: BGS_DARK
     },