+++ packages/codistica-react-icons/stories/components/linkedin.stories.js
@@ -2,21 +2,21 @@
 
 import {default as centered} from '@storybook/addon-centered/react';
 import React from 'react';
-import {BGS_DARK} from '../../../.storybook/custom-backgrounds.js';
-import {icons} from '../../../src/index.js';
+import {BGS_DARK} from '../../.storybook/custom-backgrounds.js';
+import {Linkedin as LinkedInIcon} from '../../src/index.js';
 
 /**
  * @description A Linkedin icon demo.
  * @returns {Object<string,*>} React component.
  */
 function Linkedin() {
-    return <icons.Linkedin href={'https://www.linkedin.com'} />;
+    return <LinkedInIcon href={'https://www.linkedin.com'} />;
 }
 
 export {Linkedin};
 
 export default {
-    title: 'Icons',
+    title: 'Socials',
     parameters: {
         backgrounds: BGS_DARK
     },