+++ packages/codistica-react-icons/stories/components/twitter.stories.js
@@ -2,21 +2,21 @@
 
 import {default as centered} from '@storybook/addon-centered/react';
 import React from 'react';
-import {BGS_DARK} from '../../../.storybook/custom-backgrounds.js';
-import {icons} from '../../../src/index.js';
+import {BGS_DARK} from '../../.storybook/custom-backgrounds.js';
+import {Twitter as TwitterIcon} from '../../src/index.js';
 
 /**
  * @description A Twitter icon demo.
  * @returns {Object<string,*>} React component.
  */
 function Twitter() {
-    return <icons.Twitter href={'https://twitter.com'} />;
+    return <TwitterIcon href={'https://twitter.com'} />;
 }
 
 export {Twitter};
 
 export default {
-    title: 'Icons',
+    title: 'Socials',
     parameters: {
         backgrounds: BGS_DARK
     },