+++ packages/codistica-react-icons/stories/components/youtube.stories.js
@@ -2,21 +2,21 @@
 
 import {default as centered} from '@storybook/addon-centered/react';
 import React from 'react';
-import {BGS_DARK} from '../../../.storybook/custom-backgrounds.js';
-import {icons} from '../../../src/index.js';
+import {BGS_DARK} from '../../.storybook/custom-backgrounds.js';
+import {Youtube as YoutubeIcon} from '../../src/index.js';
 
 /**
  * @description A Youtube icon demo.
  * @returns {Object<string,*>} React component.
  */
 function Youtube() {
-    return <icons.Youtube href={'https://www.youtube.com'} />;
+    return <YoutubeIcon href={'https://www.youtube.com'} />;
 }
 
 export {Youtube};
 
 export default {
-    title: 'Icons',
+    title: 'Socials',
     parameters: {
         backgrounds: BGS_DARK
     },